frontend: some fixes

+++ Tasker.Frontend/src/app/core/components/navbar/navbar.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { ConfirmationComponent } from 'src/app/shared/components/confirmation/confirmation.component';
 import { AccountService } from '../../services/account.service';
 import { TokenService } from '../../services/token.service';
 
@@ -10,12 +12,18 @@ import { TokenService } from '../../services/token.service';
 })
 export class NavbarComponent {
   constructor(public accountService: AccountService, private tokenService: TokenService,
-    private router: Router) { }
-    
+    private router: Router, private dialog: MatDialog) { }
+
   public logout(): void {
-    this.tokenService.revokeToken().subscribe(() => {
-      this.accountService.deleteCurrentUser();
-      this.router.navigate(['']);
-    });
+    this.dialog.open(ConfirmationComponent, { disableClose: true, autoFocus: false, data: { confirmationMessage: 'Confirm logout, please' } })
+      .afterClosed().subscribe((result: boolean) => {
+        if (!result) {
+          return;
+        }
+        this.tokenService.revokeToken().subscribe(() => {
+          this.accountService.deleteCurrentUser();
+          this.router.navigate(['']);
+        });
+      });
   }
 }
